Creates code architecture diagram and refactors code accordingly
