+++ ht_builder/exceptions.py
