+++ ht_builder/guess_life_aspect.py
