+++ ht_models/domain_models.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass, field
 from datetime import date
+from typing import List
 import uuid
 
 # TODO: Use a sort index for the activities that happend on the same day
@@ -20,3 +21,23 @@ class Activity:
     def print(self):
         more_info_str = "null" if self.more_info == None else f"{self.more_info}"
         return f"{self.id};{self.activity_date.strftime('%Y-%m-%d')};{self.activity_name};{self.life_aspect};{more_info_str}"
+
+
+@dataclass
+class JournalEntry:
+    record: str
+    record_date: date
+    id: uuid.UUID = field(default_factory=uuid.uuid4)
+
+    def __str__(self):
+        # TODO: Should be the decoded self.record cuz now the special characters such as '\n' are just printed as they are
+        return f"{self.record_date}\n{self.record}"
+
+    def print(self):
+        return f"{self.id};{self.record_date};{self.record}"
+
+
+@dataclass
+class UserInput:
+    activities: List[Activity]
+    journal_entries: List[JournalEntry]
