+++ setup.cfg
@@ -4,4 +4,8 @@ version = 0.0.1
 author = szabikr
 
 [options]
-packages = hparser
+packages =
+  ht_models
+  ht_parser
+  ht_builder
+  ht_importer
